Commented a little more, removed some casting warnings.
That is me done, now to prepare for my interview.

+++ cw_template/src/RigidBody.h
@@ -1,3 +1,4 @@
+// Rigid body class. This holds all the information required to simulate physics.
 #pragma once
 #include <glm/glm.hpp>
 #include "stdafx.h"
@@ -62,8 +63,8 @@ namespace phys
 		glm::dquat& GetQuat() { return orientation; }
 		glm::dvec3& GetPosition() { return position; }
 
-		void SetMass(float m) { mass = m; inverseMass = 1 / mass; }
-		float GetMass() { return mass; }
+		void SetMass(double m) { mass = m; inverseMass = 1 / mass; }
+		double GetMass() { return mass; }
 
 		void AddForceAt(const glm::dvec3 &force, const glm::dvec3 &point);
 		void AddAngularForce(const glm::dvec3 &i);
